Add 'b' option; change 'z', 'g' options; v0.2.5
- "-b" option, which allows you to read raw bytecode of
files as if they were text documents.
This still does not work with the 'u' option, so you can't
do 'sed' type stuff to directly edit image files for example.

- The "-z" option for copying files to a ZIP archive now
allows for BZIP2 and LZMA compression by calling "-zb" or "-zl",
respectively. The default "-z" is still no compression at all.

- "-sed" option for more easily searching for and replacing
patterns in text files. This is similar in concept to Linux sed.
"(<previous queries> -}})? <options> -sed '<regex>//<repl>' /dirname"
means you search all files in dirname (or in the resultset
returned by previous queries if there are any)
for <regex> and replace every instance with <repl>.
See the documentation for the "-u" option and
gorp.readfiles.GorpHandler.update(),
which this calls behind the scenes.

- The "-g" option for running a sequence of commands
in a gorp script can now be called from within a Gorp Session.
No functionality has been lost or changed.

- The is_iterable() function from gorp.utils now returns False
when called on `bytes` objects, in line with how it treats
`str` objects. This is hopefully what users would expect anyway.

- Typos in gorp.jsonpath.aggregate documentation.